[Test] Migrate all cpp tests to googletest (#2238)
* fix test_gui.cpp

* migrate ti tests to google test; fix test_gui

* [skip ci] enforce code format

* [skip ci] enforce code format
