+++ tests/cpp/common/core_test.cpp
