+++ tests/cpp/ir/algorithm_simplication_test.cpp
