+++ tests/cpp/task/test_exception_handling.cpp
